variosu tools

+++ sokker/ntdb/management/commands/copy_players_to_archive.py
@@ -2,19 +2,30 @@
 
 from django.core.management.base import BaseCommand
 from ntdb.models import Player, ArchivePlayer
-
+from django.utils.translation import gettext as _
 
 class Command(BaseCommand):
     help = "Copies records from Player table to ArchivePlayer table"
 
-    def handle(self, *args, **options):
-        # Query all records from the Player table
-        players = Player.objects.all()
+    def add_arguments(self, parser):
+        parser.add_argument(
+            '--country_code',
+            type=str,
+            help=_('Filter players by country code (e.g., 54)'),
+            required=True,
+        )
+    def handle(self, *args, **kwargs):
+        country_code = kwargs['country_code'].upper()
+        if not country_code:
+            self.stdout.write(self.style.ERROR("Country code is required"))
+            return
+        # Query players filtered by country code
+        players = Player.objects.filter(countryid=country_code).exclude(retired=True)
 
         # Copy records to ArchivePlayer table
         for player in players:
             # Get archive player by age and country
-            print(player.sokker_id, player.name, player.surname, player.age)
+            print("Archiving:",player.sokker_id, player.name, player.surname, player.age)
             aPlayer = ArchivePlayer.objects.filter(
                 sokker_id=player.sokker_id, age=player.age
             ).delete()
